+++ fileSummaryV1/service_account_keys.js
@@ -1,4 +1,5 @@
-const {bondBaseUrl, samBaseUrl, determineBondProvider, BondProviders} = require('../common/helpers');
+const {samBaseUrl} = require('../common/helpers');
+const {determineBondProvider, BondProviders} = require('../common/bond');
 const apiAdapter = require('../common/api_adapter');
 
 function maybeTalkToBond(auth, url) {
