+++ package.json
@@ -9,7 +9,7 @@
     "smoketest_v2": "ava -m smoketest_v2*",
     "smoketest_fileSummaryV1": "ava -m smoketest_fileSummaryV1*",
     "live_test": "ava -m live_test*",
-    "live_test_martha_v2": "ava test/martha_v2/martha_v2_live.test.js",
+    "live_test_martha_v2": "ava test/martha/martha_v2_live.test.js",
     "live_test_fileSummaryV1": "ava test/fileSummaryV1/fileSummaryV1_live.test.js",
     "integration": "ava -m integration*",
     "integration_v2": "ava -m integration_v2*",
