+++ test/common/helpers.test.js
@@ -1,5 +1,5 @@
 const test = require('ava');
-const {dataObjectUriToHttps, bondBaseUrl, samBaseUrl, BondProviders, determineBondProvider} = require('../../common/helpers');
+const {dataObjectUriToHttps, samBaseUrl} = require('../../common/helpers');
 const config = require('../../config.json');
 
 /**
@@ -96,44 +96,6 @@ test('dataObjectUriToHttps should throw a Error when passed an invalid uri', (t)
     }
 });
 
-test('bondBaseUrl should come from the config json', (t) => {
-    t.is(bondBaseUrl(), config.bondBaseUrl);
-});
-
 test('samBaseUrl should come from the config json', (t) => {
     t.is(samBaseUrl(), config.samBaseUrl);
-});
-
-test('BondProviders default should be "dcf-fence"', (t) => {
-    t.is(BondProviders.default, BondProviders.DCF_FENCE);
-});
-
-test('BondProviders should contain "dcf-fence" and "fence"', (t) => {
-    t.truthy(BondProviders.DCF_FENCE);
-    t.truthy(BondProviders.FENCE);
-});
-
-test('determineBondProvider should be "fence" if the URL host is "dg.4503"', (t) => {
-    t.is(determineBondProvider('drs://dg.4503/anything'), BondProviders.FENCE);
-});
-
-test('determineBondProvider should be "fence" if the URL host is "dg.712C"', (t) => {
-    t.is(determineBondProvider('drs://dg.712C/anything'), BondProviders.FENCE);
-});
-
-test('determineBondProvider should be "dcf-fence" if the URL host is "dg.foo"', (t) => {
-    t.is(determineBondProvider('drs://dg.foo/anything'), BondProviders.DCF_FENCE);
-});
-
-test('determineBondProvider should be "HCA" if the URL host ends with ".humancellatlas.org"', (t) => {
-    t.is(determineBondProvider('drs://someservice.humancellatlas.org'), BondProviders.HCA);
-});
-
-test('determineBondProvider should return the default BondProvider if the URL host does not end with ' +
-    'exactly ".humancellatlas.org"', (t) => {
-    t.is(determineBondProvider('drs://someservice.spoofhumancellatlas.org'), BondProviders.default);
-});
-
-test('determineBondProvider should return the default BondProvider if the URL host is NOT "dg.4503" or HCA', (t) => {
-    t.is(determineBondProvider('drs://some-host/anything'), BondProviders.default);
 });
